analytics-service/requirements.txt
# Analytics Service Requirements
pandas==2.1.3
numpy==1.25.2
pymongo==4.6.0
python-dateutil==2.8.2
textblob==0.17.1
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
wordcloud==1.9.2
python-dotenv==1.0.0
